int to float

+++ Odd_Even/odd_even.py
@@ -3,7 +3,7 @@
 2. use if method for selection
 """
 
-num = int(raw_input("please enter some number:>  "))
+num = float(raw_input("please enter some number:>  "))
 if (num % 2) == 0:
     #print("The number", num, "is Even")
     print("{0} is Even".format(num))
